Added user repository. Changed userId data type from int to long

+++ ChatBot.Anonymous/Domain/Entities/ChatRoom.cs
@@ -6,8 +6,8 @@ namespace ChatBot.Anonymous.Domain.Entities
     public partial class ChatRoom
     {
         public Guid Id { get; set; }
-        public int FirstUserId { get; set; }
-        public int SecondUserId { get; set; }
+        public long FirstUserId { get; set; }
+        public long SecondUserId { get; set; }
         public int NumberMessagesFirstUser { get; set; }
         public int NumberMessagesSecondUser { get; set; }
         public int StatusRoom { get; set; }
